Add files via upload
Write a game where the player has to guess a random number between 1-1000 in as few turns as possible. Use the console to print messages and take input from the player. Have a 2 option menu for the player to start a new game or exit at the start. Once playing, loop through starting new games until the player chooses to exit.